:white_check_mark: modularise coverage for role and acapy_clients
